ailurus/svcmodes/full_k8s_gcp/requirements.txt
google-auth
google-cloud-container
docker
kubernetes